Elvis test

+++ dicOcrText/normalizeImage.py
@@ -43,7 +43,9 @@ __all__ = ['getImages', 'enhanceImages']
 
 
 def getImages(images):
-    """Take input image list and save it in the scope"""
+    """
+    Take input image list and save it in the scope
+    """
     
     if len(images) >= 1 :
         num = 1
@@ -62,7 +64,7 @@ def getImages(images):
                     os.mkdir(imagedirNew)
                     
                 imagefileNew = "dic_image_"+str(num)+ext
-                imageLocationNew =  imagedirNew+'/'+imagefileNew
+                imageLocationNew =  imagedirNew+imagefileNew
                 copyfile(image, imageLocationNew)
                 MetaLex.fileImages.append(imageLocationNew)
                 num += 1
@@ -90,7 +92,9 @@ class enhanceImages ():
         self.images = MetaLex.fileImages
         
     def contrast(self, value, show=False, save=False):
-        """Enhance image file with the contrast value"""
+        """
+        Enhance image file with the contrast value
+        """
         
         if self.images >= 1 : 
             num = 1
@@ -104,11 +108,12 @@ class enhanceImages ():
                     enh.enhance(value).show()
                 elif save :
                     dicProject.createtemp()
-                    enh.enhance(value).save(tempname)
-                    dicProject.treat_image_append(tempname)
-                    message = imagename + 'is modified with contrast (' +str(value)+ ') > '+tempname+' > Saved in dictemp folder'  
-                    MetaLex.dicLog.manageLog.writelog(message) 
-                    num += 1 
+                    if dicProject.inDir(tempname) :
+                        enh.enhance(value).save(tempname)
+                        dicProject.treat_image_append(tempname)
+                        message = imagename + 'is modified with contrast (' +str(value)+ ') > '+tempname+' > Saved in dicTemp folder'  
+                        MetaLex.dicLog.manageLog.writelog(message) 
+                        num += 1 
                 else :
                     print ' Warning : contrast(value, show=False, save=False) --> You must define one action for the current treatment : show=true or save=true '
                     
@@ -119,7 +124,9 @@ class enhanceImages ():
             
             
     def sharp(self, value, show=False, save=False):
-        """Enhance image file with the sharp value"""
+        """
+        Enhance image file with the sharp value
+        """
         
         if len(self.images) >= 1 :
             num = 1
@@ -134,22 +141,25 @@ class enhanceImages ():
                     enh.enhance(value).show()
                 elif save :
                     dicProject.createtemp()
-                    enh.enhance(value).save(tempname)
-                    dicProject.treat_image_append(tempname)
-                    message = imagename + 'is modified with sharp ( ' +str(value)+ ') > '+tempname+' > Saved in dictemp folder'  
-                    MetaLex.dicLog.manageLog.writelog(message) 
-                    num += 1 
+                    if dicProject.inDir(tempname) :
+                        enh.enhance(value).save(tempname)
+                        dicProject.treat_image_append(tempname)
+                        message = imagename + 'is modified with sharp ( ' +str(value)+ ') > '+tempname+' > Saved in dicTemp folder'  
+                        MetaLex.dicLog.manageLog.writelog(message) 
+                        num += 1 
                 else :
-                    print 'Warning : sharp(value, show=False, save=False) --> You must define one action for the current treatment : show=true or save=true '
+                    print 'Warning : sharp(value, show=False, save=False) --> You must define one action for the current treatment : show=true or save=true'
         
         else:
-            message = ' Error : getImages(images) >> They are not images for the current treatement : please input images !! ' 
+            message = ' Error : getImages(images) >> They are not images for the current treatment : please input images !! ' 
             print "--> "+message+"\n"
             MetaLex.dicLog.manageLog.writelog(message)
             
             
     def bright(self, value, show=False, save=False):
-        """Enhance image file with the bright value"""
+        """
+        Enhance image file with the bright value
+        """
         
         if len(self.images) >= 1 :
             num = 1
@@ -164,50 +174,57 @@ class enhanceImages ():
                     enh.enhance(value).show()
                 elif save :
                     dicProject.createtemp()
-                    enh.enhance(value).save(tempname)
+                    if dicProject.inDir(tempname) :
+                        enh.enhance(value).save(tempname)
                     dicProject.treat_image_append(tempname)
-                    message = imagename + 'is modified with bright (' +str(value)+ ') > '+tempname+' > Saved in dictemp folder'  
+                    message = imagename + 'is modified with bright (' +str(value)+ ') > '+tempname+' > Saved in dicTemp folder'  
                     MetaLex.dicLog.manageLog.writelog(message) 
                     num += 1 
                 else :
                     print 'Warning : bright(value, show=False, save=False) --> You must define one action for the current treatment : show=true or save=true '
         else:
-            message = ' Error : getImages(images) >> They are not images for the current treatement : input images!!' 
+            message = ' Error : getImages(images) >> They are not images for the current treatment : input images!!' 
             print "--> "+message+"\n"
             MetaLex.dicLog.manageLog.writelog(message)
             
             
     def contrastBright(self, contrast, bright, show=False, save=False):
-        """Enhance image file with the contrastBright value"""
+        """
+        Enhance image file with the contrastBright value
+        """
         
         if len(self.images) >= 1 :
             num = 1
-            for image in  self.images :
-                img_conv = self.removeColor(image, save=True)
-                img = Image.open(img_conv)
+            for i, image in  enumerate(self.images) :
+                img_conv = self.removeColor(i, image, save=True)
+                imgpil = Image.open(img_conv)
                 imagename, ext = dicProject.get_part_file(image)
                 tempname = 'img_bright_'+str(num)+ext
-                enhbright = ImageEnhance.Brightness(img)
+                enhbright = ImageEnhance.Brightness(imgpil)
                 dicProject.createtemp()
                 enhbright.enhance(bright).save(tempname)
                 img2 = Image.open(tempname)
                 enhconst = ImageEnhance.Contrast(img2)
+                img_conv_part = dicProject.get_part_file(img_conv)
+                img_conv_file = img_conv_part[0]+img_conv_part[1]
                 if show :
                     enhconst.enhance(contrast).show()
                     dicProject.createtemp()
                     os.remove(tempname)
+                    os.remove(img_conv_file)
                 if save :
                     tempname2 = 'img_contrast_bright_'+str(num)+ext
                     dicProject.createtemp()
-                    enhconst.enhance(contrast).save(tempname2)
-                    os.remove(tempname)
-                    os.remove(img_conv)
-                    dicProject.treat_image_append(tempname2)
-                    
-                message = imagename + ' is modified with  contrast (' +str(contrast)+ ') and  bright ('+str(bright)+') > '+tempname2+' > Saved in dicTemp folder'  
-                MetaLex.dicLog.manageLog.writelog(message) 
-                img.close()
-                num += 1
+                    if dicProject.inDir(tempname2) :
+                        enhconst.enhance(contrast).save(tempname2)
+                        os.remove(tempname)
+                        dicProject.treat_image_append(tempname2)
+                    if dicProject.inDir(img_conv_file) :
+                        os.remove(img_conv_file)
+                    message = imagename + ' is modified with  contrast (' +str(contrast)+ ') and  bright ('+str(bright)+') > '+tempname2+' > Saved in dicTemp folder'  
+                    MetaLex.dicLog.manageLog.writelog(message) 
+                    imgpil.close()
+                    num += 1
         else:
             message = '  > They are not images for the current treatment : input images!!' 
             print "--> "+message+"\n"
@@ -215,7 +232,9 @@ class enhanceImages ():
             
            
     def convert (self, img, show=False, save=False):
-        """Convert image file to white/black image"""
+        """
+        Convert image file to white/black image
+        """
         num = 1
         if len(self.images) >= 1 :
             for image in  self.images :
@@ -226,8 +245,11 @@ class enhanceImages ():
                     img.convert("L").show()
                 if save :
                     dicProject.createtemp()
-                    img.convert("L").save(tempname)
-                    return tempname
+                    if dicProject.inDir(tempname) :
+                        img.convert("L").save(tempname)
+                        return tempname
+                    else: 
+                        return tempname
                     
         else:
             message = '  > They are not images for the current treatment : input images!!' 
@@ -236,8 +258,10 @@ class enhanceImages ():
             
                   
                 
-    def filter (self, imgfilter):
-        """Filter image file with specific filter value"""
+    def filter (self, imgfilter, show=False):
+        """
+        Filter image file with specific filter value
+        """
         
         if len(self.images) >= 1 :
             num = 1
@@ -247,9 +271,12 @@ class enhanceImages ():
                 imagename, ext = dicProject.get_part_file(image)
                 tempname = 'img_filter_'+str(num)+ext
                 dicProject.createtemp()
-                img.filter(imgfilter).save(tempname)
+                if show :
+                    img.filter(imgfilter).show()
+                elif not show and dicProject.inDir(tempname) :
+                    img.filter(imgfilter).save(tempname)
                 dicProject.treat_image_append(tempname)
-                message = imagename + ' is modified with  filter (' +str(imgfilter)+ ')  > '+tempname+' > Saved in dictemp folder'  
+                message = imagename + ' is modified with  filter (' +str(imgfilter)+ ')  > '+tempname+' > Saved in dicTemp folder'  
                 MetaLex.dicLog.manageLog.writelog(message)
                 img.close()
                 num += 1
@@ -260,35 +287,40 @@ class enhanceImages ():
             
                     
                 
-    def removeColor(self, img, show=False, save=False):
-        """Remove color in image file to enhance its quality"""
+    def removeColor(self, i, img, show=False, save=False):
+        """
+        Remove color in image file to enhance its quality
+        """
         
-        if len(self.images) >= 1 :
-            num = 1
-            for image in  self.images :
-                img = Image.open(image)
-                imagepart = dicProject.get_part_file(image)
-                tempname = 'img_color_remove_'+str(num)+imagepart[1]
-                
-                replace_color = (255, 255, 255)
-                find_color = (0, 0, 0)
-                new_image_data = []
-                for color in list(img.getdata()) :
-                    #print color
-                    if color == find_color or color == replace_color :
-                        #print color
-                        new_image_data += [color]
-                    else:
-                        #print color
-                        pass
-                        
-                img.putdata(new_image_data)
-                if show :
-                    img.show()
-                if save :
-                    dicProject.createtemp() 
-                    img.save(tempname)   
-                    return tempname
+        if img :
+            imagepart = dicProject.get_part_file(img)
+            tempname = 'img_color_remove_'+str(i)+imagepart[1]
+            
+            imgpil = Image.open(img)
+            replace_color = (255, 255, 255)
+            find_color = (0, 0, 0)
+            new_image_data = []
+            for color in list(imgpil.getdata()) :
+                if color == find_color or color == replace_color :
+                    new_image_data += [color]
+                else:
+                    pass
+            
+            imgdir = os.path.dirname(img)
+            namestore = ""
+            for tep in imgdir.split('/')[:-1] :
+                namestore += tep +"/"
+            namestore = namestore+"dicTemp/"+tempname
+            imgpil.putdata(new_image_data)
+            if show :
+                imgpil.show()
+            if save :
+                dicProject.createtemp()
+                if dicProject.inDir(tempname) : 
+                    imgpil.save(tempname) 
+                    return namestore
+                else :
+                    return namestore
         else:
             message = '  > They are not images for the current treatment : input images!!' 
             print "--> "+message+"\n"
